查看与编辑个人信息

+++ 1_Server/artwrok/src/main/java/repository/WorkRepository.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Repository
 public interface WorkRepository extends JpaRepository<WorkEntity, Integer> {
     List<WorkEntity> findByTypeId(int id);
-    List<WorkEntity> findByUid(long id);
+    List<WorkEntity> findByUid(int id);
 }
