+++ 2_Android/ArtWork/base-material/src/main/java/com/chenyuwei/basematerial/activity/BaseActivity.java
@@ -36,8 +36,8 @@ public abstract class BaseActivity extends AppCompatActivity implements View.OnC
         preferences = PreferenceManager.getDefaultSharedPreferences(activity);
     }
 
-    public long getUid(){
-        return preferences.getLong("uid",-1);
+    public int getUid(){
+        return preferences.getInt("uid",-1);
     }
 
     protected void setSupportActionBar(int id) {
